Created a SimulationExperiment view and an Issues view.
