+++ src/renderer/src/components/ContentsComponent.vue
@@ -28,57 +28,13 @@
       </Tab>
     </TabList>
     <TabPanels class="p-0!">
-      <TabPanel v-for="fileTab in fileTabs" :key="'tabPanel_' + fileTab.file.path()" :value="fileTab.file.path()">
-        <div class="tab-panel-container">
-          <Toolbar id="fileTablistToolbar" class="p-1!">
-            <template #start>
-              <Button
-                class="p-1!"
-                icon="pi pi-play-circle"
-                severity="secondary"
-                text
-                @click="onRun(fileTab as IFileTab)"
-              />
-              <Button class="p-1!" disabled icon="pi pi-stop-circle" severity="secondary" text />
-            </template>
-          </Toolbar>
-          <Splitter v-if="fileTab.file.issues().length === 0" class="border-none! h-full m-0" layout="vertical">
-            <SplitterPanel :size="89">
-              <Splitter>
-                <SplitterPanel class="ml-4 mr-4 mb-4" :size="25">
-                  <SimulationPropertyEditor :file="vue.toRaw(fileTab.file)" />
-                  <!--
-                  <SolversPropertyEditor />
-                  <GraphsPropertyEditor />
-                  <ParametersPropertyEditor />
-                  -->
-                </SplitterPanel>
-                <SplitterPanel :size="75">
-                  <v-chart autoresize :option="option" :theme="theme" />
-                </SplitterPanel>
-              </Splitter>
-            </SplitterPanel>
-            <SplitterPanel :size="11">
-              <Editor class="border-none h-full" :readonly="true" v-model="fileTab.consoleContents" />
-            </SplitterPanel>
-          </Splitter>
-          <div v-else class="issues-container">
-            <Fieldset legend="Issues">
-              <div
-                v-for="(issue, index) in fileTab.file.issues()"
-                :key="'issue_' + issue.type + '_' + issue.description"
-                :class="index > 0 ? 'mt-4!' : ''"
-              >
-                <Message v-if="issue.type === locAPI.IssueType.Error" severity="error" icon="pi pi-times-circle">
-                  {{ issue.description }}
-                </Message>
-                <Message v-else severity="warn" icon="pi pi-exclamation-triangle">
-                  {{ issue.description }}
-                </Message>
-              </div>
-            </Fieldset>
-          </div>
-        </div>
+      <TabPanel
+        v-for="(fileTab, index) in fileTabs"
+        :key="'tabPanel_' + fileTab.file.path()"
+        :value="fileTab.file.path()"
+      >
+        <SimulationExperimentView v-if="fileTab.file.issues().length === 0" v-model="fileTabs[index]" />
+        <IssuesView v-else :issues="fileTab.file.issues()" />
       </TabPanel>
     </TabPanels>
   </Tabs>
@@ -93,14 +49,7 @@ import * as locAPI from '../../../libopencor/locAPI'
 
 import * as common from '../common'
 
-import * as echarts from 'echarts/core'
-import { GridComponent, GridComponentOption, DataZoomComponent, DataZoomComponentOption } from 'echarts/components'
-import { LineChart, LineSeriesOption } from 'echarts/charts'
-import { UniversalTransition } from 'echarts/features'
-import { CanvasRenderer } from 'echarts/renderers'
-import VChart from 'vue-echarts'
-
-interface IFileTab {
+export interface IFileTab {
   file: locAPI.File
   consoleContents: string
 }
@@ -196,27 +145,9 @@ function closeAllFiles(): void {
   }
 }
 
-function onRun(fileTab: IFileTab): void {
-  const simulationTime = fileTab.file.sedInstance().run()
+// Track the height of our file tablist.
 
-  fileTab.consoleContents += `<br/>&nbsp;&nbsp;<b>Simulation time:</b> ${common.formatTime(simulationTime)}`
-
-  vue
-    .nextTick()
-    .then(() => {
-      const consoleElement = document.getElementsByClassName('ql-editor')[0]
-
-      consoleElement.scrollTop = consoleElement.scrollHeight
-    })
-    .catch((error: unknown) => {
-      console.error('Error scrolling to the bottom of the console:', error)
-    })
-}
-
-// Track our file tablist and toolbar.
-
-common.trackElementResizing('fileTablist')
-common.trackElementResizing('fileTablistToolbar')
+common.trackElementHeight('fileTablist')
 
 // Keyboard shortcuts.
 
@@ -234,135 +165,14 @@ if (!common.isMobile()) {
       event.preventDefault()
 
       selectPreviousFile()
-    } else if (!event.ctrlKey && !event.shiftKey && !event.metaKey && event.code === 'F9') {
-      event.preventDefault()
-
-      const fileTab = fileTabs.value.find((fileTab) => fileTab.file.path() === activeFile.value)
-
-      if (fileTab !== undefined) {
-        onRun(fileTab as IFileTab)
-      }
     }
   })
 }
-
-// Add an ECharts object.
-
-const theme = vue.computed(() => {
-  return common.isLightMode.value ? 'macarons' : 'dark'
-})
-
-echarts.use([GridComponent, DataZoomComponent, LineChart, CanvasRenderer, UniversalTransition])
-
-type EChartsOption = echarts.ComposeOption<GridComponentOption | DataZoomComponentOption | LineSeriesOption>
-
-function func(x: number) {
-  x /= 10
-  return Math.sin(x) * Math.cos(x * 2 + 1) * Math.sin(x * 3 + 2) * 50
-}
-
-function generateData() {
-  const data: [number, number][] = []
-
-  for (let i = -200; i <= 200; i += 0.1) {
-    data.push([i, func(i)])
-  }
-
-  return data
-}
-
-const option = vue.ref<EChartsOption>({
-  animation: false,
-  grid: {
-    top: 40,
-    left: 50,
-    right: 40,
-    bottom: 50
-  },
-  xAxis: {
-    name: 'x',
-    minorTick: {
-      show: true
-    },
-    minorSplitLine: {
-      show: true
-    }
-  },
-  yAxis: {
-    name: 'y',
-    min: -100,
-    max: 100,
-    minorTick: {
-      show: true
-    },
-    minorSplitLine: {
-      show: true
-    }
-  },
-  dataZoom: [
-    {
-      show: true,
-      type: 'inside',
-      filterMode: 'none',
-      xAxisIndex: [0],
-      startValue: -20,
-      endValue: 20
-    },
-    {
-      show: true,
-      type: 'inside',
-      filterMode: 'none',
-      yAxisIndex: [0],
-      startValue: -20,
-      endValue: 20
-    }
-  ],
-  series: [
-    {
-      type: 'line',
-      showSymbol: false,
-      clip: true,
-      data: generateData()
-    }
-  ]
-})
 </script>
 
 <style scoped>
 .file-tablist {
   border-bottom: 1px solid var(--p-primary-color);
-}
-
-.issues-container {
-  padding: var(--p-tabs-tabpanel-padding);
-}
-
-:deep(.p-button) {
-  transition: none;
-}
-
-:deep(.p-button-icon) {
-  font-size: 1.5rem;
-}
-
-:deep(.p-button-icon-only) {
-  width: 2rem;
-}
-
-:deep(.p-button-label) {
-  height: 0;
-}
-
-:deep(.ql-editor > *) {
-  cursor: default;
-}
-
-:deep(.p-editor-content) {
-  border: none !important;
-}
-
-:deep(.p-editor-toolbar) {
-  display: none;
 }
 
 .p-tab {
@@ -404,16 +214,6 @@ const option = vue.ref<EChartsOption>({
   outline: none !important;
 }
 
-.p-toolbar {
-  border: none;
-  border-radius: 0;
-  border-bottom: 1px solid var(--p-content-border-color);
-}
-
-:deep(.ql-editor) {
-  padding: 0.25rem 0.5rem;
-}
-
 .remove-button {
   padding: 0.15rem;
   font-size: 0.75rem;
@@ -429,9 +229,5 @@ const option = vue.ref<EChartsOption>({
   .remove-button:hover {
     background-color: var(--p-red-400);
   }
-}
-
-.tab-panel-container {
-  height: var(--available-viewport-height);
 }
 </style>
